Make file names fit on one line more often
This patch takes into account the file extension when truncating
file names for display, so they'll usually fit on the first line.
It feels a bit hacky, it'd be nicer to do this with css `text-overflow:
ellipsis`, but then we'd lose the file extension. If only there was a
way to force the browser to wrap anywhere, even in the middle of words,
instead of just convenient wrapping points like dashes or spaces...

This is intended to fix the wrapping issue shown in this image:

![Screen](http://i.imgur.com/qJ7osNH.png)