example: Use aliasify in example
So we can use `uppy/lib/` in the example folder.

Had to move the watch script into a JS file to do this, because aliasify
needs to be configured with a regex inside an object. We can also reuse
this watch script for future examples, although I guess we would
probably put examples in the website folder usually…

I also added `fs-write-file-atomic`, which writes to a temporary file
first and then moves it to the final `bundle.js`. This fixes an issue
where browsersync would sometimes refresh whilst browserify was still
writing to the file, and you'd get an incomplete JS file in the browser.